Here updated in getAllUsers func in auth module and rests are eslint …

+++ src/app/modules/agent/agent.repository.ts
@@ -83,9 +83,16 @@ export const AgentRepository = {
             where: { id }
         });
 
-        const user = await prisma.user.delete({
-            where: { id: agent.userId }
-        });
+        if(!agent){
+            throw new Error("Failed to delete agent");
+        }
+        else {
+            await prisma.user.delete({
+                where: {
+                    id: agent.userId
+                }
+            });
+        }
 
         return agent;
     }
